# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
      "AnswerClaim": {
        "description": "One span of the answer, and the evidence id it resolves to.\n\nThis is what makes a number in the prose clickable: the UI does not parse\nthe text looking for figures, it renders the spans the grounding gate\nalready matched.",
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "text",
          "metric_id",
          "value",
          "unit",
          "evidence_id"
        ],
        "title": "AnswerClaim",
        "type": "object"
      },
...
      "EvidenceDetail": {
        "description": "One row, its support, and everything beneath it.",
        "properties": {
          "dimension_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimension Value"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "inputs": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ]
            },
            "title": "Inputs",
            "type": "object"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "period_from": {
            "title": "Period From",
            "type": "string"
          },
          "period_to": {
            "title": "Period To",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceLevel"
          },
          "rules_applied": {
            "items": {
              "type": "string"
            },
            "title": "Rules Applied",
            "type": "array"
          },
          "source_record_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Record Ids",
            "type": "array"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          },
          "verification_checks": {
            "items": {
              "type": "string"
            },
            "title": "Verification Checks",
            "type": "array"
          }
        },
        "required": [
          "execution_id",
          "evidence_id",
          "metric_id",
          "unit",
          "value",
          "period_from",
          "period_to",
          "dimension_value",
          "inputs",
          "rules_applied",
          "verification_checks",
          "provenance",
          "source_record_ids"
        ],
        "title": "EvidenceDetail",
        "type": "object"
      },
      "EvidenceIndex": {
        "properties": {
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLine"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "execution_id",
          "items"
        ],
        "title": "EvidenceIndex",
        "type": "object"
      },
      "EvidenceLine": {
        "description": "One published metric, without its support. The index the drawer lists.",
        "properties": {
          "dimension_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimension Value"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "period_from": {
            "title": "Period From",
            "type": "string"
          },
          "period_to": {
            "title": "Period To",
            "type": "string"
          },
          "support": {
            "enum": [
              "FORMULA",
              "AGGREGATION"
            ],
            "title": "Support",
            "type": "string"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          },
          "tool_version": {
            "title": "Tool Version",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "evidence_id",
          "tool_name",
          "tool_version",
          "metric_id",
          "unit",
          "value",
          "period_from",
          "period_to",
          "dimension_value",
          "support"
        ],
        "title": "EvidenceLine",
        "type": "object"
      },
...
      "ExecutionLine": {
        "description": "One row of the history list. The question, and how it ended.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "merchant_id": {
            "title": "Merchant Id",
            "type": "string"
          },
          "period_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period From"
          },
          "period_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period To"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "response_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Source"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "execution_id",
          "merchant_id",
          "question",
          "status",
          "response_source",
          "created_at",
          "period_from",
          "period_to"
        ],
        "title": "ExecutionLine",
        "type": "object"
      },
      "ExecutionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExecutionLine"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ExecutionPage",
        "type": "object"
      },
      "ExecutionSummary": {
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/AnswerClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "execution_id": {
            "title": "Execution Id",
            "type": "string"
          },
          "grounding_attempts": {
            "title": "Grounding Attempts",
            "type": "integer"
          },
          "merchant_id": {
            "title": "Merchant Id",
            "type": "string"
          },
          "period_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period From"
          },
          "period_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period To"
          },
          "response_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Source"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "execution_id",
          "merchant_id",
          "period_from",
          "period_to",
          "status",
          "response_source",
          "answer",
          "claims",
          "grounding_attempts",
          "error"
        ],
        "title": "ExecutionSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProvenanceLevel": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "dimension_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimension Value"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "operands": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceOperand"
            },
            "title": "Operands",
            "type": "array"
          },
          "period_from": {
            "title": "Period From",
            "type": "string"
          },
          "period_to": {
            "title": "Period To",
            "type": "string"
          },
          "rules_applied": {
            "items": {
              "type": "string"
            },
            "title": "Rules Applied",
            "type": "array"
          },
          "source_record_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Record Ids",
            "type": "array"
          },
          "support": {
            "enum": [
              "FORMULA",
              "AGGREGATION"
            ],
            "title": "Support",
            "type": "string"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          },
          "tool_version": {
            "title": "Tool Version",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "evidence_id",
          "tool_name",
          "tool_version",
          "metric_id",
          "unit",
          "value",
          "period_from",
          "period_to",
          "dimension_value",
          "support",
          "detail",
          "rules_applied",
          "operands",
          "source_record_ids"
        ],
        "title": "ProvenanceLevel",
        "type": "object"
      },
      "ProvenanceOperand": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "name",
          "reference",
          "value",
          "node"
        ],
        "title": "ProvenanceOperand",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agentic financial computation and reconciliation. The LLM decides what to compute; deterministic code decides what the number is.",
    "title": "RazorMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/executions": {
      "get": {
        "description": "Newest first. What the history page renders before anything is opened.",
        "operationId": "list_history_api_v1_executions_get",
        "parameters": [
          {
            "in": "query",
            "name": "merchant_id",
            "required": true,
            "schema": {
              "title": "Merchant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List History",
        "tags": [
          "executions"
        ]
      }
    },
    "/api/v1/executions/{execution_id}": {
      "get": {
        "operationId": "get_execution_api_v1_executions__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Execution",
        "tags": [
          "executions"
        ]
      }
    },
    "/api/v1/executions/{execution_id}/evidence": {
      "get": {
        "operationId": "list_evidence_api_v1_executions__execution_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "metric_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Metric Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evidence",
        "tags": [
          "executions"
        ]
      }
    },
    "/api/v1/executions/{execution_id}/evidence/{evidence_id}": {
      "get": {
        "operationId": "get_evidence_api_v1_executions__execution_id__evidence__evidence_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "executions"
        ]
      }
    },
...
  }
}